chore: prototyping
